revert to u32

+++ src/kernel/src/fs/host/file.rs
@@ -173,7 +173,7 @@ impl HostFile {
         Ok(child)
     }
 
-    pub fn mkdir(self: &Arc<Self>, name: &str, mode: u16) -> Result<Arc<Self>, Error> {
+    pub fn mkdir(self: &Arc<Self>, name: &str, mode: u32) -> Result<Arc<Self>, Error> {
         let raw = Self::raw_mkdir(self.raw, name, mode)?;
 
         Ok(Arc::new(Self {
@@ -184,7 +184,7 @@ impl HostFile {
     }
 
     #[cfg(unix)]
-    fn raw_mkdir(parent: RawFile, name: &str, mode: u16) -> Result<RawFile, Error> {
+    fn raw_mkdir(parent: RawFile, name: &str, mode: u32) -> Result<RawFile, Error> {
         use libc::{mkdirat, mode_t};
         use std::ffi::CString;
 
@@ -198,7 +198,7 @@ impl HostFile {
     }
 
     #[cfg(windows)]
-    fn raw_mkdir(parent: RawFile, name: &str, mode: u16) -> Result<RawFile, Error> {
+    fn raw_mkdir(parent: RawFile, name: &str, mode: u32) -> Result<RawFile, Error> {
         todo!()
     }
